fix: reorg and attempt to fix CI
